[seller] refactor(173): preview 컴포넌트 로직 분리
